[rtl] move `rtllu.py` to `rtlnl.py`
