+++ yt_dlp/extractor/_extractors.py
@@ -1434,8 +1434,8 @@ from .rottentomatoes import RottenTomatoesIE
 from .rozhlas import RozhlasIE
 from .rtbf import RTBFIE
 from .rte import RteIE, RteRadioIE
-from .rtlnl import RtlNlIE
-from .rtllu import (
+from .rtlnl import (
+    RtlNlIE,
     RTLLuTeleVODIE, 
     RTLLuArticleIE,
     RTLLuTeleLiveIE,
